Added new JVM for single Projects

+++ plugins/de.ovgu.featureide.core.featurehouse.proofautomation/src/de/ovgu/featureide/core/featurehouse/proofautomation/filemanagement/FileManager.java
@@ -130,7 +130,9 @@ public class FileManager {
 		Path source = Paths.get(src);
 	    Path destination = Paths.get(target);
 	    try {
-			Files.copy(source, destination);
+	    	if(!destination.toFile().exists()){
+	    		Files.copy(source, destination);
+	    	}
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
